requirements.txt
discord.py
requests
bs4
lxml
